+++ GHEtool/VariableClasses/FluidData.py
@@ -154,3 +154,14 @@ class FluidData(BaseClass):
             if getattr(self, i) != getattr(other, i):
                 return False
         return True
+
+    def __repr__(self):
+        temp = f'Fluid parameters\n\tThermal conductivity of the fluid [W/(m·K)]: {self.k_f}\n\t' \
+               f'Density of the fluid [kg/m³]: {self.rho}\n\t' \
+               f'Thermal capacity of the fluid [J/(kg·K)]: {self.Cp}\n\t' \
+               f'Dynamic viscosity [Pa·s]: {self.mu}\n\t'
+        if self._vfr is not None:
+            temp += f'Volume flow rate [l/s]: {self.vfr}'
+        else:
+            temp += f'Mass flow rate [kg/s] : {self.mfr}'
+        return temp