+++ GHEtool/VariableClasses/LoadData/BuildingLoad/MonthlyBuildingLoadAbsolute.py
@@ -434,3 +434,22 @@ class MonthlyBuildingLoadAbsolute(_SingleYear, _LoadDataBuilding):
                 f'a length of {self.simulation_period * (8760 if self._hourly else 12)} was expected.')
 
         self._results = results
+
+    def __repr__(self):
+        temp = f'Monthly building load\n'
+        temp += f'Month\tPeak heating [kW] \t Peak cooling [kW] \t Baseload heating [kWh] \t Baseload cooling [kWh]\n'
+        for i in range(12):
+            temp += f'{i + 1}\t{self.peak_heating[i]:.2f}\t{self.peak_cooling[i]:.2f}\t' \
+                    f'{self.baseload_heating[i]:.2f}\t{self.baseload_cooling[i]:.2f}\n'
+        temp += f'Peak cooling duration [hour]: {self.peak_injection_duration / 3600:.1f}\n'
+        temp += f'Peak heating duration [hour]: {self.peak_extraction_duration / 3600:.1f}\n'
+        temp += f'Efficiency heating: {self.cop.__repr__()}\n' \
+                f'Efficiency cooling: {self.eer.__repr__()}\n' \
+                f'Simulation period [year]: {self.simulation_period}\n' \
+                f'First month of simulation [-]: {self.start_month}'
+
+        if self.max_peak_dhw == 0:
+            return temp
+
+        return temp + f'\nDHW demand [kWh/year]: {self.yearly_average_dhw_load:.0f}\n' \
+                      f'Efficiency DHW: {self.cop_dhw.__repr__()}'